requirements.txt
dj-database-url==0.4.1
Django==1.9.7
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==2.0.6
django-ckeditor==5.2.1
django-disqus==0.5
django-mathfilters==0.4.0
Pillow==4.0.0
dj-static==0.0.6